# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Platform API
  version: 2.0.0
  description: |
    基于WRF气象驱动的无人机VRP智能路径规划系统 API 平台。

    ## 功能模块
    - **平台管理**: 租户管理、API密钥管理、用量统计
    - **气象数据**: WRF/风云/天姿/风雷多源气象数据查询与融合
    - **数据同化**: 3DVAR/4DVAR/5DVAR/EnKF/Hybrid/Enhanced Bayesian 同化算法
    - **风险评估**: 综合风险感知与区域风险栅格地图
    - **适航评估**: 全维度适航评估与适航标准查询
    - **观测决策**: 自适应/计划/应急观测任务管理与决策
    - **路径规划**: VRPTW/DE-RRT*/DWA/MPC/A*/Dijkstra/RRT* 路径规划
    - **UTM**: 空域管理、飞行计划、实时追踪与冲突检测

    ## 认证方式
    - **Bearer Token**: JWT 认证，用于平台管理接口
    - **API Key**: HMAC 签名认证，用于业务接口调用

  contact:
    name: UAV Platform Team
    email: support@uav-platform.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: 开发环境（通过网关访问）
  - url: https://api.uav-platform.example.com
    description: 生产环境
...
paths:
...
  /api/v1/flight-plans:
    get:
      tags: [UTM]
      summary: 列出飞行计划
      description: 获取所有飞行计划列表。
      operationId: listFlightPlans
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListFlightPlan'

    post:
      tags: [UTM]
      summary: 提交飞行计划
      description: 提交新的飞行计划，包含无人机ID、航点列表和计划时间窗口。
      operationId: submitFlightPlan
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitFlightPlanRequest'
            example:
              uavId: "UAV-001"
              waypoints:
                - {"lon": 116.3, "lat": 39.9, "alt": 100}
                - {"lon": 116.5, "lat": 40.0, "alt": 150}
              plannedStartTime: "2025-06-01T08:00:00"
              plannedEndTime: "2025-06-01T10:00:00"
              purpose: "巡检任务"
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/flight-plans/{id}:
    get:
      tags: [UTM]
      summary: 获取飞行计划详情
      description: 根据ID获取飞行计划详细信息。
      operationId: getFlightPlan
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/flight-plans/{id}/approve:
    post:
      tags: [UTM]
      summary: 审批通过飞行计划
      description: 审批通过指定的飞行计划。
      operationId: approveFlightPlan
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 审批成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/flight-plans/{id}/reject:
    post:
      tags: [UTM]
      summary: 驳回飞行计划
      description: 驳回指定的飞行计划。
      operationId: rejectFlightPlan
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 驳回成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/flight-plans/{id}/start:
    post:
      tags: [UTM]
      summary: 开始飞行
      description: 将飞行计划状态变更为执行中。
      operationId: startFlight
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/flight-plans/{id}/complete:
    post:
      tags: [UTM]
      summary: 完成飞行
      description: 将飞行计划标记为已完成。
      operationId: completeFlight
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultFlightPlan'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # UTM - 实时追踪
  # ============================================================
  /api/v1/tracking/positions:
    post:
      tags: [UTM]
      summary: 上报无人机位置
      description: 无人机上报当前位置信息，用于实时追踪。
      operationId: reportUavPosition
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UavPositionReport'
            example:
              uavId: "UAV-001"
              lon: 116.4074
              lat: 39.9042
              alt: 100.0
              speed: 15.0
              heading: 45.0
              timestamp: "2025-06-01T08:30:00"
      responses:
        '200':
          description: 上报成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUavPosition'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/tracking/uavs/{uavId}/position:
    get:
      tags: [UTM]
      summary: 获取无人机当前位置
      description: 获取指定无人机的最新位置信息。
      operationId: getCurrentUavPosition
      security:
        - ApiKeyAuth: []
      parameters:
        - name: uavId
          in: path
          required: true
          description: 无人机ID
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUavPosition'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tracking/uavs/{uavId}/history:
    get:
      tags: [UTM]
      summary: 获取无人机轨迹历史
      description: 获取指定无人机的历史轨迹记录。
      operationId: getUavTrackHistory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: uavId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListUavPosition'
  /api/v1/tracking/conflicts/check:
    post:
      tags: [UTM]
      summary: 冲突检测
      description: 检测指定无人机在计划路径和时间窗口内是否存在飞行冲突。
      operationId: checkConflicts
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConflictCheckRequest'
            example:
              uavId: "UAV-001"
              plannedPath:
                - {"lon": 116.3, "lat": 39.9, "alt": 100, "eta": "2025-06-01T08:10:00"}
                - {"lon": 116.5, "lat": 40.0, "alt": 150, "eta": "2025-06-01T08:30:00"}
              timeWindow:
                start: "2025-06-01T08:00:00"
                end: "2025-06-01T09:00:00"
      responses:
        '200':
          description: 检测完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListConflictAlert'
        '400':
          $ref: '#/components/responses/BadRequest'

# ============================================================
# 组件定义
# ============================================================
components:
...
  schemas:
...
    Result:
      type: object
      description: 统一API响应结构
      properties:
        code:
          type: integer
          description: 业务状态码 (200=成功)
          example: 200
        message:
          type: string
          description: 提示消息
          example: success
        data:
          description: 响应数据（泛型）
        requestId:
          type: string
          description: 请求追踪ID
        timestamp:
          type: integer
          format: int64
          description: 时间戳
...
    SubmitFlightPlanRequest:
      type: object
      required: [uavId, waypoints, plannedStartTime, plannedEndTime]
      properties:
        uavId:
          type: string
          description: 无人机ID
        waypoints:
          type: array
          description: 航点列表
          items:
            $ref: '#/components/schemas/Waypoint'
        plannedStartTime:
          type: string
          format: date-time
          description: 计划开始时间
        plannedEndTime:
          type: string
          format: date-time
          description: 计划结束时间
        purpose:
          type: string
          description: 飞行目的
    Waypoint:
      type: object
      properties:
        lon:
          type: number
          format: double
        lat:
          type: number
          format: double
        alt:
          type: number
          format: double
    FlightPlan:
      type: object
      description: 飞行计划
      properties:
        id:
          type: integer
          format: int64
        uavId:
          type: string
        operatorId:
          type: string
        waypointsJson:
          type: string
          description: 航点列表JSON
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        actualStartTime:
          type: string
          format: date-time
        actualEndTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [SUBMITTED, APPROVED, REJECTED, ACTIVE, COMPLETED, CANCELLED]
          description: 飞行计划状态
        approvalCode:
          type: string
          description: 审批编号
        emergencyFlag:
          type: boolean
          description: 是否应急飞行
        tenantId:
          type: string
        createdAt:
          type: string
          format: date-time
    ResultFlightPlan:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/FlightPlan'
    ResultListFlightPlan:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FlightPlan'
    UavPositionReport:
      type: object
      required: [uavId, lon, lat, alt, timestamp]
      properties:
        uavId:
          type: string
          description: 无人机ID
        lon:
          type: number
          format: double
          description: 经度
        lat:
          type: number
          format: double
          description: 纬度
        alt:
          type: number
          format: double
          description: 海拔高度(m)
        speed:
          type: number
          format: double
          description: 速度(m/s)
        heading:
          type: number
          format: double
          description: 航向(deg)
        timestamp:
          type: string
          format: date-time
    UavPosition:
      type: object
      description: 无人机位置
      properties:
        id:
          type: integer
          format: int64
        uavId:
          type: string
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        altitude:
          type: number
          format: double
        speed:
          type: number
          format: double
        heading:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        flightPlanId:
          type: integer
          format: int64
    ResultUavPosition:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/UavPosition'
    ResultListUavPosition:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UavPosition'
    ConflictCheckRequest:
      type: object
      required: [uavId, plannedPath, timeWindow]
      properties:
        uavId:
          type: string
          description: 无人机ID
        plannedPath:
          type: array
          description: 计划路径点
          items:
            type: object
            properties:
              lon:
                type: number
                format: double
              lat:
                type: number
                format: double
              alt:
                type: number
                format: double
              eta:
                type: string
                format: date-time
        timeWindow:
          type: object
          description: 时间窗口
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
    ConflictAlert:
      type: object
      description: 冲突告警
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [GEOFENCE, UAV]
          description: 冲突类型
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          description: 严重程度
        involvedEntitiesJson:
          type: string
          description: 涉及实体JSON
        resolutionAdviceJson:
          type: string
          description: 解决建议JSON
        status:
          type: string
          enum: [ACTIVE, RESOLVED]
          description: 告警状态
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    ResultListConflictAlert:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ConflictAlert'

  # ----------------------------------------------------------
  # 安全认证
  # ----------------------------------------------------------
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，用于平台管理接口

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 认证，配合 HMAC 签名使用

  # ----------------------------------------------------------
  # 通用响应
  # ----------------------------------------------------------
  responses:
...
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1000
            message: "请求参数错误"
            data: null
            timestamp: 1717200000000
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1003
            message: "资源不存在"
            data: null
            timestamp: 1717200000000
...